refactor SandBoxApp and adding DockerSpaceViewport for testing
