+++ SandBox/imgui.ini
@@ -53,7 +53,7 @@ Size=364,202
 Collapsed=0
 
 [Window][Settings]
-Pos=3,4
+Pos=3,2
 Size=410,223
 Collapsed=1
