+++ SandBox/src/SandboxApp.cpp
@@ -4,10 +4,10 @@
 #include <MABEngine/Layers/ImGui/ImGuiLayer.h>
 #include <imgui.h>
 
-#include "SandBox3DLayer.h"
-#include "SandBox2DLayer.h"
+#include "SandBoxLayers/SandBox3DLayer.h"
+#include "SandBoxLayers/SandBox2DLayer.h"
+#include "SandBoxLayers/DockSpaceView.h"
 #include "SimpleGame/GameLayer.h"
-#include "StdViewport.h"
 
 class SandBoxApp : public MABEngine::Core::Application {
 public:
@@ -16,9 +16,10 @@ public:
 		uint32_t w = GetWindow().GetWidth();
 		uint32_t h = GetWindow().GetHeight();
 
-		//PushLayer(new SandBox3DLayer(w, h));
-		PushLayer(new SandBox2DLayer(w, h));
 		//PushLayer(new SimpleGame::GameLayer(w, h));
+		//PushLayer(new SandBoxLayers::SandBox3DLayer(w, h));
+		//PushLayer(new SandBoxLayers::SandBox2DLayer(w, h));
+		PushLayer(new SandBoxLayers::DockSpaceView(w, h));
 	}
 	~SandBoxApp() {
