--- SandBox/src/StdViewport.cpp
@@ -1,85 +1,0 @@
-#include "mabengine_pch.h"
-#include <imgui.h>
-#include <MABEngine.h>
-
-#include "StdViewport.h"
-
-
-StdViewPort::StdViewPort(uint32_t width, uint32_t height)
-	:Layer("ViewPort"), m_Width(width), m_Height(height), m_Camera(-2.0f, 2.0f, -2.0f, 2.0f)
-{
-	m_VertexArray = MABEngine::Renderer::VertexArray::Create();
-
-	float vertices[3 * 7] = {
-		-0.5f, -0.5f, 0.0f, 0.8f, 0.2f, 0.8f, 1.0f,
-		 0.5f, -0.5f, 0.0f, 0.2f, 0.1f, 0.8f, 1.0f,
-		 0.0f,  0.5f, 0.0f, 0.8f, 0.8f, 0.2f, 1.0f,
-	};
-
-	MABEngine::Core::Ref<MABEngine::Renderer::VertexBuffer> vertexBuffer;
-	vertexBuffer.reset(MABEngine::Renderer::VertexBuffer::Create(vertices, sizeof(vertices)));
-	MABEngine::Renderer::BufferLayout layout = {
-		{ MABEngine::Renderer::ShaderDataType::Float3, "a_Position"},
-		{ MABEngine::Renderer::ShaderDataType::Float4, "a_Color"},
-	};
-	vertexBuffer->SetLayout(layout);
-	m_VertexArray->AddVertexBuffer(vertexBuffer);
-
-	// Index Buffer
-	unsigned int indices[3] = { 0, 1, 2 };
-	MABEngine::Core::Ref<MABEngine::Renderer::IndexBuffer> indexBuffer;
-	indexBuffer.reset(MABEngine::Renderer::IndexBuffer::Create(indices, sizeof(indices) / sizeof(uint32_t)));
-	m_VertexArray->SetIndexBuffer(indexBuffer);
-
-	std::string vertextSrc = R"(
-			#version 330 core
-			
-			layout(location = 0) in vec3 a_Position;
-			layout(location = 1) in vec4 a_Color;
-
-			uniform	mat4 u_ViewProjection;
-
-			out vec3 v_Position;
-			out vec4 v_Color;
-
-			void main() {
-				v_Position = a_Position;
-				v_Color = a_Color;
-				gl_Position = u_ViewProjection * vec4(a_Position, 1.0);
-			}
-		)";
-
-	std::string fragmentSrc = R"(
-			#version 330 core
-			
-			layout(location = 0) out vec4 color;
-			
-			in vec3 v_Position;
-			in vec4 v_Color;
-
-			void main() {
-				color = vec4(v_Position * 0.5 + 0.5, 1.0);
-				color = v_Color;
-			}
-		)";
-
-	m_Shader = MABEngine::Renderer::Shader::Create("testPackage", vertextSrc, fragmentSrc);
-}
-
-StdViewPort::~StdViewPort()
-{
-
-}
-
-void StdViewPort::OnImGuiRender()
-{
-	ImGui::Begin("Viewport");
-	// Set up a viewport or sub-window where OpenGL will draw the 3D box
-	ImVec2 viewportPanelSize = ImGui::GetContentRegionAvail();
-
-	if (viewportPanelSize.x > 0 && viewportPanelSize.y > 0) {
-	}
-
-
-	ImGui::End();
-}
